Implemented searching of songs Fixed view of the songs

+++ system/player.php
@@ -11,10 +11,25 @@ function recursive_unset(&$array, $unwanted_key) {
     }
 }
 
-function search_song($id) {
-    GLOBAL $db;
+function search_songs($search) {
+    global $db;
+    $songs = array();
 
     foreach($db as $arrID => $data) {
+        if ((strpos($data["name"], $search) !== false)) {
+            array_push($songs, $data);
+        } else if ((strpos($data["artist"], $search) !== false)) {
+            array_push($songs, $data);
+        }
+    }
+
+    return $songs;
+}
+
+function search_song($id) {
+    global $db;
+
+    foreach ($db as $arrID => $data) {
         if ($data["id"] == $id)
             return $data;
     }
@@ -27,7 +42,7 @@ if (isset($_GET["id"])) {
     include_once __DIR__ . "/vendor/autoload.php";
 
     $song = search_song($_GET["id"]);
-    $newName =  "/temp/" . bin2hex(random_bytes(22)) . ".mp3";
+    $newName = "/temp/" . bin2hex(random_bytes(22)) . ".mp3";
 
     $time = 0;
     $timeSet = false;
@@ -51,10 +66,14 @@ if (isset($_GET["id"])) {
     echo json_encode($song);
 } else {
     recursive_unset($db, "fileName");
-    echo json_encode($db);
-}
-$size = ob_get_length();
 
+    if (isset($_GET["search"]) && $_GET["search"] !== "")
+        echo json_encode(search_songs($_GET["search"]));
+    else
+        echo json_encode($db);
+}
+
+$size = ob_get_length();
 header("Content-Encoding: none");
 header("Content-Length: " . $size);
 header("Connection: close");
@@ -63,13 +82,10 @@ ob_end_flush();
 @ob_flush();
 flush();
 
-if(session_id()) session_write_close();
+if (session_id()) session_write_close();
 
 $files = glob(__DIR__ . "/temp/*");
-
 foreach ($files as $file) {
-    if (is_file($file)) {
-        if (time() - filemtime($file) >= 60 * 60 * 24)
-            unlink($file);
-    }
+    if (is_file($file) && time() - filemtime($file) >= 60 * 60 * 24)
+        unlink($file);
 }
